requirements.txt
pydantic
python-dotenv

django

langchain
langchain_google_genai
langchain_core
langgraph


google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
google-cloud-aiplatform

numpy
requests
pathlib